refactor: readme
